POC; migrating java tests to dsmTest
